refactor: 微调代码&fix：eslint
