+++ .eslintrc.cjs
@@ -1,7 +1,4 @@
 module.exports = {
   root: true,
   extends: ['@element-plus/eslint-config'],
-  rules: {
-    'vue/no-mutating-props': 'off'
-  }
 }
